refactor(zim): extract _StructureMixin to zim/structure.py
Move article-structure / link / TOC methods (get_article_structure,
extract_article_links, get_table_of_contents, _build_hierarchical_toc,
_headings_to_tree, get_related_articles, _resolve_link_to_entry_path
and the _extract_* helpers) into _StructureMixin.

ZimOperations now inherits from _SearchMixin, _ContentMixin and
_StructureMixin.
